+++ server/homebrew.model.js
@@ -4,11 +4,12 @@ const _ = require('lodash');
 const zlib = require('zlib');
 
 const HomebrewSchema = mongoose.Schema({
-	shareId : { type: String, default: ()=>{return nanoid(12);}, index: { unique: true } },
-	editId  : { type: String, default: ()=>{return nanoid(12);}, index: { unique: true } },
-	title   : { type: String, default: '' },
-	text    : { type: String, default: '' },
-	textBin : { type: Buffer },
+	shareId   : { type: String, default: ()=>{return nanoid(12);}, index: { unique: true } },
+	editId    : { type: String, default: ()=>{return nanoid(12);}, index: { unique: true } },
+	title     : { type: String, default: '' },
+	text      : { type: String, default: '' },
+	textBin   : { type: Buffer },
+	pageCount : { type: Number, default: 1 },
 
 	description : { type: String, default: '' },
 	tags        : { type: String, default: '' },